requirements.txt
fastapi
asyncpg
uvicorn
pydantic
tortoise-orm
passlib
bcrypt
pyjwt
python-multipart